requirements.txt
requests
python-dotenv
openpyxl
fpdf2
